requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
SpeechRecognition==3.10.0
pydub==0.25.1
redis==5.0.1
tenacity==8.2.3
PyPDF2==3.0.1
python-docx==1.1.0
pytest==7.4.0
pytest-asyncio==0.21.0